+++ scs-client/C#/SCSSdkClient/Object/SCSTelemetry.cs
@@ -45,10 +45,30 @@ namespace SCSSdkClient.Object {
         public bool SdkActive { get; internal set; }
 
         /// <summary>
-        ///     Timestamp not the in game time,
+        ///      Similar to simulation time however it stops
+        ///      when the physics simulation is paused.
+        ///      Is not reseted.
         /// </summary>
         /// <seealso cref="Common.GameTime" />
-        public uint Timestamp { get; internal set; }
+        /// <seealso cref="SimulationTimestamp" />
+        public ulong Timestamp { get; internal set; }
+
+        /// <summary>
+        ///     Time controlling the physical simulation.
+        ///
+        ///     Usually changes with fixed size steps so it oscilates
+        ///     around the render time. This value changes even if the
+        ///     physics simulation is currently paused.
+        /// </summary>
+        /// <seealso cref="RenderTimestamp" />
+        public ulong SimulationTimestamp { get; internal set; }
+
+        /// <summary>
+        ///     Time controlling the visualization.
+        ///
+        ///     Its step changes depending on rendering FPS.
+        /// </summary>
+        public ulong RenderTimestamp { get; internal set; }
 
         /// <summary>
         ///     Is the game actual paused? (Menu, F1, Map, etc. open)